requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
fal_client==0.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
pillow==10.4.0
pyTelegramBotAPI==4.21.0
requests==2.32.3
sniffio==1.3.1
telebot==0.0.5
urllib3==2.2.2
